Create view lastValues.jsp.
This page is the boundary between the client and LastValueControl. The expected
data model from this control is a Map<String,Object>, where the key is the name
of the variable or the timestamp and the value is the double value of the variable or
a Date instance that represents the timestamp.

The timestamp is formatted according to this pattern: "dd/MM/yyyy hh:mm:ss aa".

The response of this view is JSON and the JSON Tag Library is used to create it. This
tag library can be found here: <http://json-taglib.sourceforge.net/>.