Modify view index.jsp to add variable and measures fields dynamically.
The variable and measures fields are now added according to the listings DaoList
and MeasuresList. These classes are dummy classes and should be merged with
the measure and dao branches.

The file /web/WEB-INF/config.properties contains the display names and measu-
rement units of the variables and the names of the measures.